Use f64 instead of u32 for file positions (#78)

+++ src/wasm/bgzf_wasm.d.ts
@@ -5,8 +5,8 @@ export class ChunkSliceResult {
   private constructor();
   free(): void;
   [Symbol.dispose](): void;
-  readonly cpositions: Uint32Array;
-  readonly dpositions: Uint32Array;
+  readonly cpositions: Float64Array;
+  readonly dpositions: Float64Array;
   readonly buffer: Uint8Array;
 }
 
@@ -22,4 +22,8 @@ export function decompress_all(input: Uint8Array): Uint8Array;
 
 export function decompress_block(input: Uint8Array): DecompressResult;
 
+/**
+ * Decompress a slice of BGZF data between two virtual offsets.
+ * Position parameters use f64 to map to JS number, supporting files >4GB.
+ */
 export function decompress_chunk_slice(input: Uint8Array, min_block_position: number, min_data_position: number, max_block_position: number, max_data_position: number): ChunkSliceResult;
